+++ src/main/java/com/chattriggers/ctjs/internal/mixins/ChatComponentMixin.java
@@ -1,5 +1,6 @@
 package com.chattriggers.ctjs.internal.mixins;
 
+import com.chattriggers.ctjs.api.client.Client;
 import com.chattriggers.ctjs.api.message.ChatLib;
 import net.minecraft.client.gui.components.ChatComponent;
 import net.minecraft.client.multiplayer.chat.GuiMessage;
@@ -11,6 +12,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.List;
@@ -20,6 +22,16 @@ public class ChatComponentMixin {
     @Final
     @Shadow
     private List<GuiMessage> allMessages;
+
+    @ModifyVariable(
+        method = "addMessage(Lnet/minecraft/network/chat/Component;Lnet/minecraft/network/chat/MessageSignature;Lnet/minecraft/client/multiplayer/chat/GuiMessageSource;Lnet/minecraft/client/multiplayer/chat/GuiMessageTag;)V",
+        at = @At("HEAD"),
+        ordinal = 0,
+        argsOnly = true
+    )
+    private Component v5$addMessage(Component original) {
+        return Client.processName(original);
+    }
 
     @Inject(method = "clearMessages", at = @At("TAIL"))
     private void injectClear(boolean clearHistory, CallbackInfo ci) {
