+++ src/main/java/com/chattriggers/ctjs/internal/mixins/ConnectionProxyMixin.java
@@ -1,4 +1,4 @@
-package com.v5.mixins;
+package com.chattriggers.ctjs.internal.mixins;
 
 import com.chattriggers.ctjs.api.client.Proxy;
 import com.chattriggers.ctjs.api.client.ProxyInfo;
