+++ src/main/java/com/chattriggers/ctjs/internal/mixins/GuiRendererMixin.java
@@ -1,7 +1,7 @@
-package com.v5.mixins;
+package com.chattriggers.ctjs.internal.mixins;
 
-import com.mojang.blaze3d.buffers.GpuBufferSlice;
 import com.chattriggers.ctjs.api.render.NVGRenderer;
+import com.mojang.blaze3d.buffers.GpuBufferSlice;
 import net.minecraft.client.gui.render.GuiRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -9,14 +9,14 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(GuiRenderer.class)
-public class GameRendererMixin {
+public class GuiRendererMixin {
     @Inject(method = "render", at = @At("HEAD"))
     private void beforeRender(GpuBufferSlice fogBuffer, CallbackInfo ci) {
         NVGRenderer.runPreDrawables();
     }
 
     @Inject(method = "render", at = @At("TAIL"))
-    private void onRender(GpuBufferSlice fogBuffer, CallbackInfo ci) {
+    private void afterRender(GpuBufferSlice fogBuffer, CallbackInfo ci) {
         NVGRenderer.runDrawables();
     }
 }
