--- src/main/java/com/v5/mixins/BlockMixin.java
@@ -1,39 +1,0 @@
-package com.v5.mixins;
-
-import com.chattriggers.ctjs.api.client.Client;
-import java.util.Set;
-import net.minecraft.world.level.block.Block;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(Block.class)
-public class BlockMixin {
-    private static final Set<String> TARGET_KEYS =
-            Set.of(
-                    "block.minecraft.melon",
-                    "block.minecraft.pumpkin",
-                    "block.minecraft.carrots",
-                    "block.minecraft.potatoes",
-                    "block.minecraft.wheat",
-                    "block.minecraft.nether_wart",
-                    "block.minecraft.sugar_cane",
-                    "block.minecraft.cactus",
-                    "block.minecraft.cocoa",
-                    "block.minecraft.melon_stem",
-                    "block.minecraft.pumpkin_stem",
-                    "block.minecraft.carved_pumpkin");
-
-    @Inject(method = "spawnDestroyParticles", at = @At("HEAD"), cancellable = true)
-    private void v5$spawnBreakParticles(CallbackInfo ci) {
-        if (!Client.getHideParticles()) {
-            return;
-        }
-
-        String blockKey = ((Block) (Object) this).getDescriptionId();
-        if (TARGET_KEYS.contains(blockKey)) {
-            ci.cancel();
-        }
-    }
-}
