--- src/main/java/com/v5/mixins/ChatComponentMixin.java
@@ -1,20 +1,0 @@
-package com.v5.mixins;
-
-import com.chattriggers.ctjs.api.client.Client;
-import net.minecraft.client.gui.components.ChatComponent;
-import net.minecraft.network.chat.Component;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
-
-@Mixin(ChatComponent.class)
-public class ChatComponentMixin {
-    @ModifyVariable(
-            method = "addMessage(Lnet/minecraft/network/chat/Component;Lnet/minecraft/network/chat/MessageSignature;Lnet/minecraft/client/multiplayer/chat/GuiMessageSource;Lnet/minecraft/client/multiplayer/chat/GuiMessageTag;)V",
-            at = @At("HEAD"),
-            ordinal = 0,
-            argsOnly = true)
-    private Component v5$addMessage(Component original) {
-        return Client.processName(original);
-    }
-}
