+++ src/main/resources/ctjs.mixins.json
@@ -9,6 +9,7 @@
         "BlockEntityRenderStateMixin",
       "BookViewScreenAccessor",
       "BossHealthOverlayAccessor",
+      "CameraMixin",
       "ChatComponentAccessor",
       "ChatComponentMixin",
         "ChatScreenAccessor",
@@ -20,16 +21,26 @@
       "MultiPlayerGameModeMixin",
       "ClientPacketListenerAccessor",
       "ClientLevelAccessor",
+      "ConnectionProxyMixin",
       "CreativeModeInventoryScreenMixin",
+      "CrossCollisionBlockMixin",
+        "EntityRendererMixin",
         "EntityRenderDispatcherMixin",
         "EntityRenderStateMixin",
+      "FirmamentModAnnouncerMixin",
+      "FramerateLimitTrackerMixin",
+      "FrustumMixin",
       "OptionsAccessor",
       "OptionsMixin",
       "AbstractContainerScreenAccessor",
       "AbstractContainerScreenMixin",
       "GuiMixin",
+      "GuiRendererMixin",
+      "ItemInHandRendererMixin",
+      "JoinMultiplayerScreenMixin",
       "KeyMappingAccessor",
         "KeyMappingAccessor$Category",
+      "KeyboardHandlerMixin",
       "MinecraftAccessor",
       "MinecraftMixin",
       "MouseHandlerMixin",
@@ -40,6 +51,8 @@
       "PlayerTabOverlayMixin",
         "RenderTickCounterMixin",
       "LevelRendererMixin",
+      "VisGraphMixin",
+      "YggdrasilServicesKeyInfoMixin",
       "commands.ClientSuggestionProviderMixin",
       "commands.ClientPacketListenerMixin",
         "sound.SoundAccessor",
