--- src/main/resources/v5.mixins.json
@@ -1,34 +1,0 @@
-{
-    "required": true,
-    "minVersion": "0.8",
-    "package": "com.v5.mixins",
-    "compatibilityLevel": "JAVA_25",
-    "client": [
-        "BlockMixin",
-        "CameraMixin",
-        "ChatComponentMixin",
-        "ConnectionProxyMixin",
-        "CrossCollisionBlockMixin",
-        "DrawContextCapture",
-        "EntityRendererMixin",
-        "FirmamentModAnnouncerMixin",
-        "FramerateLimitTrackerMixin",
-        "FrustumMixin",
-        "GameRendererMixin",
-        "ItemInHandRendererMixin",
-        "KeyboardHandlerMixin",
-        "MinecraftMixin",
-        "MouseHandlerMixin",
-        "JoinMultiplayerScreenMixin",
-        "PlayerTabOverlayMixin",
-        "LevelRendererMixin",
-        "VisGraphMixin",
-        "YggdrasilServicesKeyInfoMixin"
-    ],
-    "mixins": [
-        "ConnectionMixin"
-    ],
-    "injectors": {
-        "defaultRequire": 1
-    }
-}
